Conversation Callables from the PSR-11 Container
This enables the use of services names as the callables for conversation answers.

So if you had a service like this in your PSR 11 container:
```php
class ReceiveEmailResponse {
    public function __construct(NewsletterManager $newsletterManager)
    {
        $this->newsletterManager = newsletterManager;
    }
    public function __invoke($answer, $conversation)
    {
        $email = $answer->getText();
        $this->newsletterManager->sendNewsletter($email)
    }
}
```
and have added that container to the bot:
```php
$botman->setContainer($psr11Container);
$botman->hears('Hello', function($bot) {
    $bot->startConversation(new OnboardingConversation);
});
```
you can then do this.
```php
    class OnboardingConversation extends Conversation {
        public function run() {
            $this->ask("What is your email address?", ReceiveEmailResponse::class)
        }
    }
```

It has bonuses of separating concerns, reusability and I've noticed that Xdebug is a lot happier debugging code this way than using closures.

Still to do: tests for all other ask methods.